issue fixes #63 audit all activities on each microservice

+++ db/oracle/create_oracle.sql
@@ -78,7 +78,7 @@ CREATE TABLE refresh_token (
 );
 
 create table audit_log (
-  log_id INT, -- system milliseonds from 1970.
+  log_id numeric, -- system milliseonds from 1970.
   service_id VARCHAR2(32) NOT NULL,
   endpoint VARCHAR2(256) NOT NULL,
   request_header VARCHAR2(4000),
